같은 날짜 같은 식물에 물주기 중복 기록 방지, 오늘 미룸/화분 안마름 정보 뿌리기, 모음 판별 메소드 및 json 키값 수정

+++ front-end/src/components/Modal/DeleteModal.jsx
@@ -37,11 +37,8 @@ const DeleteModal = ({title, url, path, deleteTooltipMsg, deleteCallBackFunction
     await deleteData(`${url}/${path}`);
     closeDeleteModal();
 
-    if (deleteCallBackFunction) {
-      deleteCallBackFunction();
-    }
+    deleteCallBackFunction && deleteCallBackFunction();
 
-    // TODO state delete?
     navigate(url, {replace: true, state: "delete"});
   };
